fix compile c++ with dfsan bug

+++ llvm_mode/libcxx_dfsan/compile.sh
@@ -1,5 +1,6 @@
 #!/usr/bin/env bash
 
+LLVM_VERSION=4.0.0
 
 NINJA_B=`which ninja 2>/dev/null`
 
@@ -10,47 +11,55 @@ if [ "$NINJA_B" = "" ]; then
 fi
 
 set -euxo pipefail
-wget http://releases.llvm.org/4.0.1/llvm-4.0.1.src.tar.xz
-wget http://releases.llvm.org/4.0.1/cfe-4.0.1.src.tar.xz
-wget http://releases.llvm.org/4.0.1/compiler-rt-4.0.1.src.tar.xz
-wget http://releases.llvm.org/4.0.1/libcxx-4.0.1.src.tar.xz
-wget http://releases.llvm.org/4.0.1/libcxxabi-4.0.1.src.tar.xz
-wget http://releases.llvm.org/4.0.1/libunwind-4.0.1.src.tar.xz
-wget http://releases.llvm.org/4.0.1/clang-tools-extra-4.0.1.src.tar.xz
 
 CUR_DIR=`pwd`
 CLANG_SRC=${CUR_DIR}/llvm_src
+
+if [ ! -d $CLANG_SRC ]; then
+wget http://releases.llvm.org/${LLVM_VERSION}/llvm-${LLVM_VERSION}.src.tar.xz
+wget http://releases.llvm.org/${LLVM_VERSION}/cfe-${LLVM_VERSION}.src.tar.xz
+wget http://releases.llvm.org/${LLVM_VERSION}/compiler-rt-${LLVM_VERSION}.src.tar.xz
+wget http://releases.llvm.org/${LLVM_VERSION}/libcxx-${LLVM_VERSION}.src.tar.xz
+wget http://releases.llvm.org/${LLVM_VERSION}/libcxxabi-${LLVM_VERSION}.src.tar.xz
+wget http://releases.llvm.org/${LLVM_VERSION}/libunwind-${LLVM_VERSION}.src.tar.xz
+wget http://releases.llvm.org/${LLVM_VERSION}/clang-tools-extra-${LLVM_VERSION}.src.tar.xz
+
+
 rm -rf $CLANG_SRC
 
-tar -Jxf ${CUR_DIR}/llvm-4.0.1.src.tar.xz 
-mv llvm-4.0.1.src $CLANG_SRC
+tar -Jxf ${CUR_DIR}/llvm-${LLVM_VERSION}.src.tar.xz 
+mv llvm-${LLVM_VERSION}.src $CLANG_SRC
 
 cd ${CLANG_SRC}/tools
-tar -Jxf ${CUR_DIR}/cfe-4.0.1.src.tar.xz 
-mv cfe-4.0.1.src clang
+tar -Jxf ${CUR_DIR}/cfe-${LLVM_VERSION}.src.tar.xz 
+mv cfe-${LLVM_VERSION}.src clang
 cd ${CLANG_SRC}/tools/clang/tools
-tar -Jxf ${CUR_DIR}/clang-tools-extra-4.0.1.src.tar.xz 
-mv clang-tools-extra-4.0.1.src extra
+tar -Jxf ${CUR_DIR}/clang-tools-extra-${LLVM_VERSION}.src.tar.xz 
+mv clang-tools-extra-${LLVM_VERSION}.src extra
 cd ${CLANG_SRC}/projects
-tar -Jxvf ${CUR_DIR}/compiler-rt-4.0.1.src.tar.xz
-mv compiler-rt-4.0.1.src compiler-rt
-tar -Jxvf ${CUR_DIR}/libcxx-4.0.1.src.tar.xz
-mv libcxx-4.0.1.src libcxx
-tar -Jxvf ${CUR_DIR}/libcxxabi-4.0.1.src.tar.xz
-mv libcxxabi-4.0.1.src libcxxabi
-tar -Jxvf ${CUR_DIR}/libunwind-4.0.1.src.tar.xz
-mv libunwind-4.0.1.src libunwind
+tar -Jxvf ${CUR_DIR}/compiler-rt-${LLVM_VERSION}.src.tar.xz
+mv compiler-rt-${LLVM_VERSION}.src compiler-rt
+tar -Jxvf ${CUR_DIR}/libcxx-${LLVM_VERSION}.src.tar.xz
+mv libcxx-${LLVM_VERSION}.src libcxx
+tar -Jxvf ${CUR_DIR}/libcxxabi-${LLVM_VERSION}.src.tar.xz
+mv libcxxabi-${LLVM_VERSION}.src libcxxabi
+tar -Jxvf ${CUR_DIR}/libunwind-${LLVM_VERSION}.src.tar.xz
+mv libunwind-${LLVM_VERSION}.src libunwind
+cp ./libcxxabi/include/*  ./libcxx/include
 
 rm -rf ${CUR_DIR}/*.tar.xz
+fi
 
 cd $CUR_DIR
+
+rm -rf build
 mkdir build && cd build/
 
-CC=~/angora/bin/angora-clang CXX=~/angora/bin/angora-clang++ cmake -G Ninja ../llvm_src  -DLIBCXXABI_ENABLE_SHARED=NO -DLIBCXX_ENABLE_SHARED=NO -DLLVM_FORCE_USE_OLD_TOOLCHAIN=YES -DLIBCXX_CXX_ABI=libcxxabi
-
+CC=~/angora/bin/angora-clang CXX=~/angora/bin/angora-clang++ cmake -G Ninja ../llvm_src  -DLIBCXXABI_ENABLE_SHARED=NO -DLIBCXX_ENABLE_SHARED=NO -DLIBCXX_CXX_ABI=libcxxabi 
+#-DLLVM_FORCE_USE_OLD_TOOLCHAIN=YES 
 USE_DFSAN=1 ninja cxx cxxabi
 
-@echo "if cxxabi.h not found, try: cp ./libcxxabi/include/*  ./libcxx/include, or -I"
+# @echo "if cxxabi.h not found, try: cp ./libcxxabi/include/*  ./libcxx/include, or -I"
 
 cp $CUR_DIR/build/lib/libc++.a $CUR_DIR/.
 cp $CUR_DIR/build/lib/libc++abi.a $CUR_DIR/libc++abidfsan.a
